+++ receiver/src/statistics/statistics_sender_influx_db.cpp
@@ -10,7 +10,7 @@ namespace asapo {
 
 template<typename ... Args>
 std::string string_format( const std::string& format, Args ... args ) {
-    size_t size = snprintf( nullptr, 0, format.c_str(), args ... ) + 1;
+    size_t size = static_cast<size_t>(snprintf(nullptr, 0, format.c_str(), args ...) + 1);
     std::unique_ptr<char[]> buf( new char[ size ] );
     snprintf( buf.get(), size, format.c_str(), args ... );
     return std::string( buf.get(), buf.get() + size - 1 );
