+++ tests/automatic/system_io/ip_tcp_network/client_serv_multicon/multicon.cpp
@@ -86,7 +86,7 @@ void CheckNormal(int times) {
     for (int i = 0; i < times; i++) {
         std::cout << "[CLIENT] send random number" << std::endl;
 
-        uint64_t message_send = rand();
+        uint64_t message_send = static_cast<uint64_t>(rand());
 
         std::cout << "[CLIENT] Send Size" << std::endl;
         io->Send(socket, &message_send, sizeof(uint64_t), &err);
