+++ tests/manual/producer_cpp/producer.cpp
@@ -38,10 +38,10 @@ int main(int argc, char* argv[]) {
 
 
     if(argc >= 2)
-        submodule = atoi(argv[1]);
+        submodule = static_cast<uint32_t>(atoi(argv[1]));
 
     if(argc >= 3)
-        sleeptime = atoi(argv[2]);
+        sleeptime = static_cast<uint32_t>(atoi(argv[2]));
 
 
     asapo::Error err;
